Remove comparer overriding when caching table objects.
We don't want to change the behaviour of the underlying table. If the
source database has a case-sensitive comparer we don't want that to make
it case-insensitive as that could cause objects whose name only differ
in case to be excluded.

Continue the comparer overriding behaviour for the relational database
layer however. This is because if we cache based on a case-insensitive
comparer, then this enables easier querying for most databases in
addition to improving performance by caching. This can cause collisions
but only if we choose to do so.