Remove remnants of dependencies and expressions.
Dependencies and dependents are not used yet. Additionally the way they
were presented was too simplistic to be useful. This is to be revisited
at a later point.

Expressions are another example where they were too simplistic to be
useful. The dependencies were never really used properly outside of the
reflection/modelled scope so it was moved from Core to Reflection.Model.