Add initial implementation of a transparent caching layer.
It is currently commented out due to limitations from corefx. The
intention is for the connection to automatically cache all results from
the database in the form of a DataTable.

The advantage of this approach is that duplicated queries for async/sync
are automatically cached and that the database layer doesn't need to
know about caching at all. Should provide a large performance
improvement, particularly to high latency connections.