Large refactoring of core interfaces. Ensures data relationships are uni-directional instead of bi-directional.
Primarily this affects objects that relate to tables. For example, a unique key no longer has a reference to it's parent table. This simplifies object creation and more easily enables serialization. Because the root/parent object is required to access the objects, the reference was merely a convenience in the first place.

* Triggers, indexes, keys no longer have a parent reference.
* Relational keys now have a reference by name to the parent/child tables.
* Specialization of some interfaces into Table/View versions is no longer necessary (e.g. IDatabaseTableColumn) and has been removed.
* IDependentRelationalDatabase has been removed, it was always clumsy and has not been implemented anyway.