Now ignoring server and database names in database layer queries.
This is because a different connection would be required to get different results anyway. Could alternatively error out if there is a mismatch.

Choosing not to error out because we want queries to be simple, for the same reason that a default schema is implied on a query for an object.