First round of migration to constraint based assertions.
i.e. Assert.That(...) for tests instead of the "classic" model.

The motivation for this is primarily for improved error messages. However, the additional consistency and clarity is nice too. For example, some of the collection based tests can now be asserted declaratively, instead of just using something like Assert.True/False() on the result of a check.